pyproject.toml
[project]
name = "tracecolor"
dynamic = ["version"]
authors = [
  { name="Marco Del Pin", email="marco.delpin@gmail.com" },
]
description = "Enhanced Python logger with colorized output, TRACE/PROGRESS levels, UDP monitoring, file logging, and external config support via Loguru backend."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "loguru>=0.7.2",
    "colorlog>=6.0.0",
]
license = { file = "LICENSE" }

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
]

[project.scripts]
tracecolor-monitor = "tracecolor.monitor:main"

[project.urls]
Homepage = "https://github.com/marcodelpin/tracecolor"
Issues = "https://github.com/marcodelpin/tracecolor/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "tracecolor.__version__"}

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3